Use aligned loads instead of unaligned
With our compiler flags, `_mm256_loadu_si256` generates two dependent
instructions:

    ```asm
    vmovdqu xmm5, XMMWORD PTR [rdi+r12]
    vinserti128     ymm1, ymm5, XMMWORD PTR [rdi+16+r12], 0x1
    ```

These instructions are a bottleneck on the critical path for stage 1,
taking up roughly 20% of the 70-80 cycle long iteration.

They each take 6-7 cycles and are *dependent* on each other, blocking
anything that uses the JSON input (which is understandably
quite a lot, since this is a JSON parser) for 13 cycles. The extra
instruction size (which we double because we load 2 256-bit registers)
takes an extra cycle to decode, meaning it'll be another cycle before
we can even try to fill the time with unrelated work.

Switching to an aligned load, `_mm256_load_si256`, generates a single
instruction with half the latency:

    ```asm
    vmovdqa ymm1, YMMWORD PTR [rdi+r12]
    ```

It requires that input be aligned.