[WIP] Backport to ARMv7-A.
With some small polyfills to simulate the new Q-form instructions with
2 D-form instructions, we can add support for ARMv7-A and cover basically
the entire mobile market.

Makefile flags need to be updated, but tests seem to be passing on a
Snapdragon 835 in 32-bit mode.

Additionally, I'm not sure if we should change the paths and namespaces
to NEON or ARM or whatever. However, since the code is technically
optimized for ARM64 and is backported to ARMv7-A, it is probably okay.