Remove flags for static/shared and libc++
These are something that must be controlled from outside, because they
are not build nor usage requirements of this library. Reference:
https://cmake.org/cmake/help/v3.12/variable/BUILD_SHARED_LIBS.html

But how do you build just this library as shared in a bigger build tree?
You can use a special variable that includes a cmake module at project()
call, which can contain a set() call to set the value of
BUILD_SHARED_LIBS at the library's directory scope only. Reference:
https://cmake.org/cmake/help/v3.12/variable/CMAKE_PROJECT_PROJECT-NAME_INCLUDE.html

For the libc++ flags you can do the same as above. Since it seems to be
an optional part of the project, I see no reason to have it as a switch.
If, for example, this library used the Coroutines TS from the libc++
standard library, then it of course needs to be part of the lists files,
as that is an obvious build requirement.